+++ base/math.jl
@@ -929,9 +929,10 @@ mod2pi(x) = rem2pi(x,RoundDown)
 """
     muladd(x, y, z)
 
-Combined multiply-add, computes `x*y+z` allowing the add and multiply to be contracted with
-each other or ones from other `muladd` and `@fastmath` to form `fma`
-if the transformation can improve performance.
+Combined multiply-add: computes `x*y+z`, but allowing the add and multiply to be merged
+with each other or with surrounding operations for performance.
+For example, this may be implemented as an [`fma`](@ref) if the hardware supports it
+efficiently.
 The result can be different on different machines and can also be different on the same machine
 due to constant propagation or other optimizations.
 See [`fma`](@ref).