+++ stdlib/LinearAlgebra/test/generic.jl
@@ -238,6 +238,7 @@ end
     @test !issymmetric(NaN)
 end
 
+@test rank(fill(0, 0, 0)) == 0
 @test rank([1.0 0.0; 0.0 0.9],0.95) == 1
 @test qr(big.([0 1; 0 0]))[2] == [0 1; 0 0]