+++ stdlib/Markdown/test/runtests.jl
@@ -1096,3 +1096,11 @@ let
     md.meta[:foo] = 42
     @test !haskey(md′.meta, :foo)
 end
+
+let
+    v = Markdown.parse("foo\n\n- 1\n- 2\n\n- 3\n\n\n- 1\n- 2\n\nbar\n\n- 1\n\n  2\n- 4\n\nbuz\n\n- 1\n- 2\n  3\n- 4\n")
+    @test v.content[2].loose
+    @test !v.content[3].loose
+    @test v.content[5].loose
+    @test !v.content[7].loose
+end