+++ stdlib/Pkg3/src/API.jl
@@ -74,7 +74,7 @@ function up(ctx::Context, pkgs::Vector{PackageSpec};
     else
         for reg in registries()
             if isdir(joinpath(reg, ".git"))
-                regpath = pathrepr(reg)
+                regpath = pathrepr(ctx, reg)
                 printpkgstyle(ctx, :Updating, "registry at ", regpath)
                 LibGit2.with(LibGit2.GitRepo, reg) do repo
                     if LibGit2.isdirty(repo)